@@ -1,27 +1,34 @@
 [build-system]
-requires = ["setuptools>=61"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling>=1.18"]
+build-backend = "hatchling.build"
 
 [project]
 name = "static_error_handler"
-version = "0.1.4"
+dynamic = ["version"]
 description = "A minimal Python adaptation of Rust's Result type"
 readme = "README.md"
-authors = [{name = "J-joon"}]
-license = {file = "LICENSE"}
-requires-python = ">=3.10"
+authors = [{ name = "J-joon" }]
+license = { file = "LICENSE" }
+
+requires-python = ">=3.13"
+
 classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent",
 ]
 
-[tool.setuptools]
-package-dir = {"" = "src"}
+[tool.hatch.version]
+path = "src/static_error_handler/__init__.py"
 
-[tool.setuptools.packages.find]
-where = ["src"]
+[tool.hatch.build.targets.wheel]
+packages = ["src/static_error_handler"]
+include = ["src/static_error_handler/py.typed"]
 
-[tool.setuptools.package-data]
-static_error_handler = ["py.typed"]
-
+[tool.hatch.build.targets.sdist]
+include = [
+  "src/static_error_handler/**",
+  "README.md",
+  "LICENSE",
+  "pyproject.toml",
+]
