@@ -1,36 +1,52 @@
-from __future__ import annotations
-"""Typed implementation of a minimal Result type."""
-
-__version__ = "0.1.4"
-
-from typing import TypeVar, Callable, Generic, NoReturn, TypeAlias
 from dataclasses import dataclass
-
-T = TypeVar("T")
-U = TypeVar("U")
-E = TypeVar("E")
-F = TypeVar("F")
+from typing import Callable, NoReturn, Protocol, runtime_checkable
 
 
 def panic(message: str) -> NoReturn:
-    """
-    Raises a RuntimeError with the given message, mimicking Rust's panic! macro.
-
-    Args:
-        message (str): The error message to display.
-
-    Raises:
-        RuntimeError: Always raised with the provided message.
-    """
     raise RuntimeError(f"Panic: {message}")
 
-class Result(Generic[T, E]):
-    """Base class for Ok and Err to define the Result type."""
-    # This class should probably be abstract and have no direct implementation
-    pass
-    
+
+@runtime_checkable
+class Result[T, E](Protocol):
+    # Predicates
+    def is_ok(self) -> bool: ...
+    def is_ok_and(self, op: Callable[[T], bool]) -> bool: ...
+    def is_err(self) -> bool: ...
+    def is_err_and(self, op: Callable[[E], bool]) -> bool: ...
+
+    # Extractors
+    def ok(self) -> T | None: ...
+    def err(self) -> E | None: ...
+
+    # Transforms
+    def map[U](self, op: Callable[[T], U]) -> "Result[U, E]": ...
+    def map_or[U](self, default: U, op: Callable[[T], U]) -> U: ...
+    def map_or_else[U](self, default: Callable[[E], U], op: Callable[[T], U]) -> U: ...
+    def map_err[F](self, op: Callable[[E], F]) -> "Result[T, F]": ...
+
+    # Unwrap / expect
+    def expect(self, msg: str) -> T: ...
+    def unwrap(self) -> T: ...
+    def expect_err(self, msg: str) -> E: ...
+    def unwrap_err(self) -> E: ...
+
+    # Combinators
+    def and_[U](self, res: "Result[U, E]") -> "Result[U, E]": ...
+    def and_then[U](self, op: Callable[[T], "Result[U, E]"]) -> "Result[U, E]": ...
+    def or_[F](self, res: "Result[T, F]") -> "Result[T, F]": ...
+    def or_else[F](self, op: Callable[[E], "Result[T, F]"]) -> "Result[T, F]": ...
+
+    # Defaults
+    def unwrap_or(self, default: T) -> T: ...
+    def unwrap_or_else(self, op: Callable[[E], T]) -> T: ...
+
+    # Side effects
+    def inspect(self, op: Callable[[T], None]) -> "Result[T, E]": ...
+    def inspect_err(self, op: Callable[[E], None]) -> "Result[T, E]": ...
+
+
 @dataclass
-class Ok(Result[T, E]):
+class Ok[T, E](Result[T, E]):
     value: T
 
     def is_ok(self) -> bool:
@@ -51,16 +67,16 @@ class Ok(Result[T, E]):
     def err(self) -> E | None:
         return None
 
-    def map(self, op: Callable[[T], U]) -> Result[U, E]:
+    def map[U](self, op: Callable[[T], U]) -> Result[U, E]:
         return Ok(op(self.value))
 
-    def map_or(self, default: U, op: Callable[[T], U]) -> U:
+    def map_or[U](self, default: U, op: Callable[[T], U]) -> U:
         return op(self.value)
 
-    def map_or_else(self, default: Callable[[E], U], op: Callable[[T], U]) -> U:
+    def map_or_else[U](self, default: Callable[[E], U], op: Callable[[T], U]) -> U:
         return op(self.value)
 
-    def map_err(self, op: Callable[[E], F]) -> Result[T, F]:
+    def map_err[F](self, op: Callable[[E], F]) -> Result[T, F]:
         return Ok(self.value)
 
     def expect(self, msg: str) -> T:
@@ -75,16 +91,16 @@ class Ok(Result[T, E]):
     def unwrap_err(self) -> E:
         panic("unwrap_err")
 
-    def and_(self, res: Result[U, E]) -> Result[U, E]:
+    def and_[U](self, res: Result[U, E]) -> Result[U, E]:
         return res
 
-    def and_then(self, op: Callable[[T], Result[U, E]]) -> Result[U, E]:
+    def and_then[U](self, op: Callable[[T], Result[U, E]]) -> Result[U, E]:
         return op(self.value)
 
-    def or_(self, res: Result[T, F]) -> Result[T, F]:
+    def or_[F](self, res: Result[T, F]) -> Result[T, F]:
         return Ok(self.value)
 
-    def or_else(self, op: Callable[[E], Result[T, F]]) -> Result[T, F]:
+    def or_else[F](self, op: Callable[[E], Result[T, F]]) -> Result[T, F]:
         return Ok(self.value)
 
     def unwrap_or(self, default: T) -> T:
@@ -102,7 +118,7 @@ class Ok(Result[T, E]):
 
 
 @dataclass
-class Err(Result[T, E]):
+class Err[T, E](Result[T, E]):
     error: E
 
     def is_ok(self) -> bool:
@@ -123,16 +139,16 @@ class Err(Result[T, E]):
     def err(self) -> E | None:
         return self.error
 
-    def map(self, op: Callable[[T], U]) -> Result[U, E]:
+    def map[U](self, op: Callable[[T], U]) -> Result[U, E]:
         return Err(self.error)
 
-    def map_or(self, default: U, op: Callable[[T], U]) -> U:
+    def map_or[U](self, default: U, op: Callable[[T], U]) -> U:
         return default
 
-    def map_or_else(self, default: Callable[[E], U], op: Callable[[T], U]) -> U:
+    def map_or_else[U](self, default: Callable[[E], U], op: Callable[[T], U]) -> U:
         return default(self.error)
 
-    def map_err(self, op: Callable[[E], F]) -> Result[T, F]:
+    def map_err[F](self, op: Callable[[E], F]) -> Result[T, F]:
         return Err(op(self.error))
 
     def expect(self, msg: str) -> T:
@@ -147,16 +163,16 @@ class Err(Result[T, E]):
     def unwrap_err(self) -> E:
         return self.error
 
-    def and_(self, res: Result[U, E]) -> Result[U, E]:
+    def and_[U](self, res: Result[U, E]) -> Result[U, E]:
         return Err(self.error)
 
-    def and_then(self, op: Callable[[T], Result[U, E]]) -> Result[U, E]:
+    def and_then[U](self, op: Callable[[T], Result[U, E]]) -> Result[U, E]:
         return Err(self.error)
 
-    def or_(self, res: Result[T, F]) -> Result[T, F]:
+    def or_[F](self, res: Result[T, F]) -> Result[T, F]:
         return res
 
-    def or_else(self, op: Callable[[E], Result[T, F]]) -> Result[T, F]:
+    def or_else[F](self, op: Callable[[E], Result[T, F]]) -> Result[T, F]:
         return op(self.error)
 
     def unwrap_or(self, default: T) -> T:
@@ -172,10 +188,5 @@ class Err(Result[T, E]):
         op(self.error)
         return self
 
-__all__ = [
-    "Ok",
-    "Err",
-    "Result",
-    "panic",
-    "__version__",
-]
+
+__all__ = ["Ok", "Err", "Result", "panic", "__version__"]
