@@ -1,3 +1,4 @@
-from .static_error_handler import Ok, Err, Result, panic, __version__
+from .static_error_handler import Ok, Err, Result, panic
 
 __all__ = ["Ok", "Err", "Result", "panic", "__version__"]
+__version__ = "0.1.4"
